requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==4.0
django-allauth==0.47.0
django-cors-headers==3.10.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
fonttools==4.28.5
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
oauthlib==3.1.1
opencv-python==4.5.5.62
packaging==21.3
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tweepy==4.4.0
tzdata==2021.5
urllib3==1.26.7
wget==3.2
